+++ src/Model/Component/Cache/Exception/NativeCacheExceptionInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace YamlStandards\Model\Component\Cache\Exception;
+
+interface NativeCacheExceptionInterface
+{
+}