+++ src/Model/YamlIndent/YamlIndentFixer.php
@@ -17,9 +17,9 @@ class YamlIndentFixer extends AbstractFixer
     /**
      * @inheritDoc
      */
-    public function fix(string $pathToYamlFile, string $pathToDumpFixedFile, StandardParametersData $standardParametersData): Result
+    public function fix(string $pathToFile, string $pathToDumpFixedFile, StandardParametersData $standardParametersData): Result
     {
-        $fileContent = file_get_contents($pathToYamlFile);
+        $fileContent = file_get_contents($pathToFile);
         $fileContent = str_replace("\r", '', $fileContent); // remove carriage returns
         $fileLines = explode("\n", $fileContent);
         $yamlIndentDataFactory = new YamlIndentDataFactory();
@@ -32,7 +32,7 @@ class YamlIndentFixer extends AbstractFixer
         $rightFileContent = implode("\n", $rightFileLines);
 
         if ($fileContent === $rightFileContent) {
-            return new Result($pathToYamlFile, Result::RESULT_CODE_OK);
+            return new Result($pathToFile, Result::RESULT_CODE_OK);
         }
 
         file_put_contents($pathToDumpFixedFile, $rightFileContent);
@@ -40,6 +40,6 @@ class YamlIndentFixer extends AbstractFixer
         $differ = new Differ();
         $diffBetweenStrings = $differ->diff($fileContent, $rightFileContent);
 
-        return new Result($pathToYamlFile, Result::RESULT_CODE_FIXED_INVALID_FILE_SYNTAX, $diffBetweenStrings);
+        return new Result($pathToFile, Result::RESULT_CODE_FIXED_INVALID_FILE_SYNTAX, $diffBetweenStrings);
     }
 }